requirements.txt
aniso8601==3.0.0
argon2-cffi==18.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
peewee==3.3.3
pycparser==2.18
pytz==2018.4
six==1.11.0
Werkzeug==0.14.1
WTForms==2.1
